Move RuntimeInformation to dedicated header
