+++ src/index/PatternCreator.h
@@ -17,6 +17,7 @@
 #include "index/StxxlSortFunctors.h"
 #include "util/BufferedVector.h"
 #include "util/ExceptionHandling.h"
+#include "util/HashMap.h"
 #include "util/Serializer/SerializeVector.h"
 #include "util/Serializer/Serializer.h"
 #include "util/TypeTraits.h"
